+++ lib/kettle/dev/template_helpers.rb
@@ -345,6 +345,8 @@ module Kettle
           c = c.gsub(/\bKettle::Dev\b/u, namespace) unless namespace.empty?
           c = c.gsub("Kettle%3A%3ADev", namespace_shield) unless namespace_shield.empty?
           c = c.gsub("kettle--dev", gem_shield)
+          # Replace require and path structures with gem_name, modifying - to / if needed
+          c = c.gsub("kettle/dev", gem_name.tr("-", "/"))
         end
         c
       end