+++ CHANGELOG.md
@@ -35,10 +35,11 @@ Please file a bug if you notice a violation of semantic versioning.
 - 76.23% documented
 ### Added
 - kettle:dev:install can overwrite gemspec with example gemspec
-- documentation for start_step CLI option for kettle-release
+- documentation for the start_step CLI option for kettle-release
 - kettle:dev:install add `only=` option for glob filtering.
 ### Fixed
 - kettle:dev:install remove "Works with MRI Ruby*" lines with no badges left
+- kettle:dev:install prefix badge cell replacement with a single space
 
 ## [1.0.17] - 2025-08-29
 - TAG: [v1.0.17][1.0.17t]