+++ CHANGELOG.md
@@ -22,8 +22,15 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Deprecated
 ### Removed
 ### Fixed
-- replacement logic handles a dashed gem-name which maps onto a nested path structure
 ### Security
+
+## [1.0.19] - 2025-08-29
+- TAG: [v1.0.19][1.0.19t]
+- COVERAGE: 97.84% -- 2350/2402 lines in 20 files
+- BRANCH COVERAGE: 81.46% -- 962/1181 branches in 20 files
+- 76.23% documented
+### Fixed
+- replacement logic handles a dashed gem-name which maps onto a nested path structure
 
 ## [Unreleased]
 ### Added
@@ -291,7 +298,7 @@ Please file a bug if you notice a violation of semantic versioning.
   - Selecting will run the selected workflow via `act`
   - This may move to its own gem in the future.
 
-[Unreleased]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.18...HEAD
+[Unreleased]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.19...HEAD
 [1.0.0]: https://github.com/kettle-rb/kettle-dev/compare/a427c302df09cfe4253a7c8d400333f9a4c1a208...v1.0.0
 [1.0.0t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.0
 [1.0.1]: https://gitlab.com/kettle-rb/kettle-dev/-/compare/v1.0.0...v1.0.1
@@ -330,3 +337,5 @@ Please file a bug if you notice a violation of semantic versioning.
 [1.0.17t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.17
 [1.0.18]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.17...v1.0.18
 [1.0.18t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.18
+[1.0.19]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.18...v1.0.19
+[1.0.19t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.19